Assured web application is still functional
