+++ messaging/src/test/java/com/ultimatesoftware/workflow/messaging/correlation/TenantUtilsTest.java
@@ -4,6 +4,8 @@ import static com.ultimatesoftware.workflow.messaging.Constants.ZERO_UUID;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.UUID;
+
+import com.ultimatesoftware.workflow.messaging.TenantUtils;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
