+++ webapp/src/main/java/com/ultimatesoftware/workflow/webapp/CustomSpringProcessEngineConfiguration.java
@@ -9,5 +9,4 @@ public class CustomSpringProcessEngineConfiguration extends SpringProcessEngineC
         this.bpmnParseFactory = bpmnParseFactory;
     }
 
-
 }
