+++ webapp/src/main/java/com/ultimatesoftware/workflow/webapp/WebappApplication.java
@@ -6,9 +6,8 @@ import com.jayway.jsonpath.spi.json.JacksonJsonNodeJsonProvider;
 import com.jayway.jsonpath.spi.json.JsonProvider;
 import com.jayway.jsonpath.spi.mapper.JacksonMappingProvider;
 import com.jayway.jsonpath.spi.mapper.MappingProvider;
-import com.ultimatesoftware.workflow.messaging.consumer.TopicContainerManager;
 import com.ultimatesoftware.workflow.messaging.kafka.CorrelatingMessageListener;
-import com.ultimatesoftware.workflow.messaging.kafka.KafkaUtils;
+import com.ultimatesoftware.workflow.messaging.kafka.TopicContainerManager;
 import com.ultimatesoftware.workflow.messaging.topicmapping.MessageTypeMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
@@ -65,7 +64,7 @@ public class WebappApplication implements ApplicationListener<ApplicationReadyEv
     @Override
     public void onApplicationEvent(ApplicationReadyEvent applicationReadyEvent) {
         topicContainerManager.createOrStartConsumers(
-                KafkaUtils.getTopics(messageTypeMapper),
+                TopicUtil.getTopics(messageTypeMapper),
                 listener);
     }
